+++ R/chains.R
@@ -1,13 +1,14 @@
-#' Extract Chain-Codes from a PDB Object
+#' Extract Chain-Codes from a PDB Molecule
 
 #' Extract the codes of all chains from a PDB object.
 
 #' \code{chains} is a generic function to extract the ids
-#' of the chains from a PDB object.
+#' of the chains from a PDB molecule.
 
 
 #' @param x Object of class \code{pdb} or class \code{atoms};
-#' @param \dots arguments passed to methods.
+#' @param sorted logical indicating if the returned chain-codes should be sorted;
+#' @param \dots arguments passed to other methods.
 
 #' @return \code{chains} returns the codes of the individual chains in the molecule.